Move the class NavigatorByTypeModel to top level and have it maintained by SingleModelGuiElements class. Move SingleModelGuiElements so it is constructed on main thread and not in an observer to be accessible from OpenSimDB. Fix references accordingly
