+++ Gui/opensim/plotter/src/org/opensim/plotter/PlotterQuantityNameFilterJPanel.java
@@ -366,7 +366,7 @@ public class PlotterQuantityNameFilterJPanel extends javax.swing.JPanel
         JComboBox cb = (JComboBox)evt.getSource();
         currentModel = (Model) cb.getSelectedItem();
         // Find model's muscle groups and fill in drop down'
-        Vector<String> groups=ViewDB.getInstance().getModelGuiElements(currentModel).getActuatorGroupNames();
+        Vector<String> groups=OpenSimDB.getInstance().getModelGuiElements(currentModel).getActuatorGroupNames();
         jMuscleGroupComboBox.setModel(new DefaultComboBoxModel(groups));
 
     }//GEN-LAST:event_jModelComboBoxActionPerformed
@@ -408,7 +408,7 @@ public class PlotterQuantityNameFilterJPanel extends javax.swing.JPanel
         jModelComboBox.setModel(new DefaultComboBoxModel(models));
         currentModel = OpenSimDB.getInstance().getCurrentModel();
         jModelComboBox.setSelectedItem(currentModel);
-        Vector<String> groups=ViewDB.getInstance().getModelGuiElements(currentModel).getActuatorGroupNames();
+        Vector<String> groups=OpenSimDB.getInstance().getModelGuiElements(currentModel).getActuatorGroupNames();
         jMuscleGroupComboBox.setModel(new DefaultComboBoxModel(groups));
         setFilter(currentFilter);
         if (groups.size()>0){
@@ -446,7 +446,7 @@ public class PlotterQuantityNameFilterJPanel extends javax.swing.JPanel
          // update names form model, muscle goup selection
          String muscleGroup = (String)jMuscleGroupComboBox.getSelectedItem();
          Model model = (Model)jModelComboBox.getSelectedItem();
-         Vector<String> names = ViewDB.getInstance().getModelGuiElements(model).getActuatorNamesForGroup(muscleGroup);
+         Vector<String> names = OpenSimDB.getInstance().getModelGuiElements(model).getActuatorNamesForGroup(muscleGroup);
          tableModel.markSelectedNames(names);
       }
       updateSelected();
