+++ Gui/opensim/view/src/org/opensim/view/markerEditor/OneMarkerDeleteAction.java
@@ -91,7 +91,7 @@ public final class OneMarkerDeleteAction extends CallableSystemAction {
         markerset.remove(marker);
 
         // Update the marker name list in the ViewDB.
-        ViewDB.getInstance().getModelGuiElements(model).updateMarkerNames();
+        OpenSimDB.getInstance().getModelGuiElements(model).updateMarkerNames();
 
         // Generate an event so everyone can update, including the marker editor.
         Vector<OpenSimObject> objs = new Vector<OpenSimObject>(1);
