+++ Gui/opensim/view/src/org/opensim/view/nodes/ConnectionEditor.java
@@ -79,7 +79,7 @@ public class ConnectionEditor {
         else 
             ViewDB.getInstance().updateModelDisplay(model);
         if (node!= null) node.refreshNode();
-        SingleModelGuiElements guiElem = ViewDB.getInstance().getModelGuiElements(model);
+        SingleModelGuiElements guiElem = OpenSimDB.getInstance().getModelGuiElements(model);
         guiElem.setUnsavedChangesFlag(true);
     }
